+++ qucs-filter/qf_math.h
@@ -2,7 +2,7 @@
 #define QF_MATH
 
 namespace qf {
-typedef double qf_float;
+typedef long double qf_float;
 static const double pi = 3.1415926535897932384626433832795029; /* pi       */
 static const double one_over_pi = 0.3183098861837906715377675267450287;                       /* 1/pi     */
 static const double ln2 = 0.6931471805599453094172321214581766; /* log_e(2) */
